% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cub.heatmap.r
\name{cub.heatmap}
\alias{cub.heatmap}
\title{CUB Heatmap}
\usage{
cub.heatmap <- function(cub.list , path.save = "./" , cellwidth = 45 , 
                  cellheight = 30 , fontsize = 25 , height = 22 , 
                  width = 30 , clustering_distance_rows = "correlation" , 
                  clustering_distance_cols = "correlation" , 
                  clustering_method = "ward.D")
}
\arguments{
\item{cub.list}{A list of dinucleotide or codon usage.}

\item{path.save}{A file path where to save the picture.}

\item{cellwidth}{An individual cell width in points. If left as NA, then the values depend on the size of plotting window. Default is 45}

\item{cellheight}{An individual cell height in points. If left as NA, then the values depend on the size of plotting window. Default is 30}

\item{fontsize}{A base fontsize for the plot. Default is 25}

\item{height}{A manual option for determining the output file height in inches. Default is 22}

\item{width}{manual option for determining the output file width in inches. Default is 30}

\item{clustering_distance_rows}{distance measure used in clustering rows. Possible values are "correlation" for Pearson correlation and all the distances supported by dist, such as "euclidean", etc. If the value is none of the above it is assumed that a distance matrix is provided. Default is euclidean}

\item{clustering_distance_cols}{distance measure used in clustering columns. Possible values the same as for clustering_distance_rows. Default is euclidean}

\item{clustering_method}{clustering method used. Accepts the same values as hclust. Default is ward.D}

\item{draw}{Draw Heatmap for dinucleotide then draw should be = "di" Default. if RSCU then draw should be = "rscu"}
}
\value{
Invisibly a pheatmap object
}
\description{
Make a GC box plot
}
\examples{
\dontrun{

# read DNA from fasta file
fasta.virus <- read.virus(c("test1.fasta", "test2.fasta", "test3.fasta"))
# Calculate zscore using (base model)
base <- dinuc.base(fasta.virus, permutations = 100)
cub.heatmap(base)
}
}
\author{
Ali Mostafa Anwar \email{aliali.mostafa99@gmail.com} , Salma Bayoumi {salma.ismail.hamed@gmail.com}
}
