% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinuc.base.R
\name{dinuc.base}
\alias{dinuc.base}
\title{Statistical dinucleotide over- and underrepresentation (base model).}
\usage{
dinuc.base(df.virus,permutations=100,exact_numbers = FALSE)
}
\arguments{
\item{df.virus}{a list returned from read.virus function}

\item{permutations}{the number of permutations for the z-score computation.  Default is 100}

\item{exact_numbers}{if TRUE exact analytical calculation will be used.  Default is FALSE}
}
\value{
A list of data.frame(s) containing the computed statistic for each dinucleotide in all DNA sequences.
}
\description{
A measure of statistical dinucleotide over- and underrepresentation; by allows for random sequence generation by shuffling (with/without replacement) of all bases in the sequence.
}
\details{
For more information \href{https://www.rdocumentation.org/packages/seqinr/versions/3.6-1/topics/dinucleotides}{seqinr}.
}
\note{
If there are NAs values in the data it will by imputed by  1/5 the lowest value in the data.
}
\examples{
\dontrun{
# read DNA from fasta file
fasta.virus <- read.virus(c("test1.fasta", "test2.fasta", "test3.fasta"))
# Calculate zscore using (base model)
base <- dinuc.base(fasta.virus, permutations = 100)
}
}
\author{
Ali Mostafa Anwar \email{aliali.mostafa99@gmail.com} , Salma Bayoumi {salma.ismail.hamed@gmail.com}
}
