% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinuc.codon.R
\name{dinuc.codon}
\alias{dinuc.codon}
\title{Statistical dinucleotide over- and underrepresentation (codon model).}
\usage{
dinuc.codon(df.virus,permutations=100,exact_numbers = FALSE)
}
\arguments{
\item{df.virus}{a list returned from read.virus function}

\item{permutations}{the number of permutations for the z-score computation.  Default is 100}

\item{exact_numbers}{if TRUE exact analytical calculation will be used.  Default is FALSE}
}
\value{
A list of data.frame(s) containing the computed statistic for each dinucleotide in all DNA sequences within df.virus.
}
\description{
A measure of statistical dinucleotide over- and underrepresentation; by allows for random sequence generation by shuffling (with/without replacement) of codons.
}
\details{
For more information \href{https://www.rdocumentation.org/packages/seqinr/versions/3.6-1/topics/dinucleotides}{seqinr}.
}
\examples{
\dontrun{
# read DNA from fasta file
fasta.virus <- read.virus(c("test1.fasta", "test2.fasta", "test3.fasta"))
# Calculate zscore using (base model)
base <- dinuc.codon(fasta.virus, permutations = 100)
}
}
\author{
Ali Mostafa Anwar \email{aliali.mostafa99@gmail.com} , Salma Bayoumi {salma.ismail.hamed@gmail.com}
}
