% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENc.GC3plot.group.R
\name{ENc.GC3plot.group}
\alias{ENc.GC3plot.group}
\title{ENc-GC3 scatterplot.}
\usage{
ENc.GC3plot.group(enc.list, gc.list)
}
\arguments{
\item{enc.list}{a list of ENc contains all the data frame with ENc values.}

\item{gc.df}{a list of ENc contains all the data frame with GC3 values.}
}
\value{
A ggplot object.
}
\description{
Make an ENc-GC3 scatterplot. Where the y-axis represents the ENc values and the x-axis represents the GC3 content.
The red fitting line shows the expected ENc values when codon usage bias affected solely by GC3.
}
\details{
For more information about ENc-GC3 plot \href{https://www.tandfonline.com/doi/full/10.1038/emi.2016.106}{Butt et al., 2016}.
}
\examples{
\dontrun{
# read DNA from fasta file
df <- read.virus(c("test1.fasta", "test2.fasta", "test3.fasta") , sep = "_")

GC.list <- GC.content(df)
enc.novembre <- ENc.values.new(df)

ENc.GC3plot.group(enc.novembre , GC.list)
}
}
\author{
Ali Mostafa Anwar \email{aliali.mostafa99@gmail.com} , Salma Bayoumi {salma.ismail.hamed@gmail.com}
}
