% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENc.values.old.R
\name{ENc.values.old}
\alias{ENc.values.old}
\title{Effective Number of Codons (ENc).}
\usage{
ENc.values.old <- function(df.fasta, genetic.code = "1", threshold = 0, alt.init = TRUE , stop.rm = FALSE , filtering = "none")
}
\arguments{
\item{df.fasta}{A list returned from read.virus function or read.host function}

\item{genetic.code}{A single string that uniquely identifies the genetic code to extract. Should be one of the values in the id or name2 columns of GENETIC_CODE_TABLE.}

\item{threshold}{Optional numeric, specifying sequence length, in codons, used for filtering}

\item{alt.init}{Logical, whether to use alternative initiation codons. Default is TRUE.}

\item{stop.rm}{Logical, whether to remove stop codons. Default is FALSE.}

\item{filtering}{Character vector, one of c("none", "soft", "hard"). Specifies whether sequences shorther than some threshold value of length (in codons), len.threshold, should be excluded from calculations. If "none" (default), length of sequences is not checked, if "soft", a warrning is printed if there are shorter sequences, and if "hard", these sequences are excluded from calculation.}
}
\value{
A list of data.frame(s) containing the computed ENc values for each DNA sequences.
}
\description{
Measure the Effective Number of Codons (ENc) of DNA sequence. Using version Wright (1990).
}
\details{
For more information about ENc \href{https://pubmed.ncbi.nlm.nih.gov/2110097/}{Wright, 1990}.
}
\examples{
\dontrun{
# read DNA from fasta file
fasta.virus <- read.virus(c("test1.fasta", "test2.fasta", "test3.fasta"))
# Calculate enc version Wright (1990)
enc.wright <- ENc.values.old(fasta.virus)
}
}
\author{
Ali Mostafa Anwar \email{aliali.mostafa99@gmail.com} , Salma Bayoumi {salma.ismail.hamed@gmail.com}
}
