% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC.boxplot.r
\name{GC.boxplot}
\alias{GC.boxplot}
\title{GC box plot}
\usage{
GC.boxplot(fasta.list)
}
\arguments{
\item{fasta.list}{A list of GC content from GC.content function.}
}
\value{
A ggplot object.
}
\description{
Make a GC box plot
}
\examples{
\dontrun{

# read DNA from fasta file
fasta.virus <- read.virus(c("test1.fasta", "test2.fasta", "test3.fasta"))
# Calculate GC content
gc.df <- GC.content(fasta.virus)
GC.boxplot(gc.df)

}
}
\author{
Ali Mostafa Anwar \email{aliali.mostafa99@gmail.com} , Salma Bayoumi {salma.ismail.hamed@gmail.com}
}
