% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC.content.R
\name{GC.content}
\alias{GC.content}
\title{GC content}
\usage{
GC.content(df.virus)
}
\arguments{
\item{df.virus}{a list returned from read.virus function}
}
\value{
A list of data.frame(s) with overall GC content as well as GC at first, second, and third codon positions of all DNA sequence.
}
\description{
Calculates overall GC content as well as GC at first, second, and third codon positions.
}
\examples{
\dontrun{
# read DNA from fasta file
fasta.virus <- read.virus(c("test1.fasta", "test2.fasta", "test3.fasta"))
# Calculate GC content
gc.df <- GC.content(fasta.virus)
}
}
\author{
Ali Mostafa Anwar \email{aliali.mostafa99@gmail.com} , Salma Bayoumi {salma.ismail.hamed@gmail.com}
}
