% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neutrality.plot.r
\name{Neutrality.plot}
\alias{Neutrality.plot}
\title{Neutrality plot plot}
\usage{
Neutrality.plot(gc.df, size = 5)
}
\arguments{
\item{gc.df}{A dataframe contains the GC12 and GC3 calculated using the GC.content() function.}
}
\value{
A ggplot object.
}
\description{
Neutrality plots of GC12 in function of GC3 was generated to assess the extent of mutational pressure on the usage of codons A correlation among GC12 and GC3 indicates mutational forces.
}
\details{
For more information about Neutrality plot \href{https://doi.org/10.1016/j.meegid.2020.104471}{Nambou.K and Anakpa.M, 2020}.
}
\examples{
\dontrun{
# read DNA from fasta file
df <- read.virus(c("test1.fasta", "test2.fasta", "test3.fasta"))
GC.list <- GC.content(df)
Neutrality.plot(GC.list[["test1"]], size = 5)
}

}
\author{
Ali Mostafa Anwar \email{aliali.mostafa99@gmail.com} , Salma Bayoumi {salma.ismail.hamed@gmail.com}
}
