% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PR2.plot.R
\name{PR2.plot}
\alias{PR2.plot}
\title{Parity rule 2 (PR2) plot}
\usage{
PR2.plot(fasta.df , fold4 = TRUE)
}
\arguments{
\item{fasta.df}{a DNAStringSet object contain only one data frame.}

\item{fold4}{if TRUE only 4-fold amino acids will be included in the analysis.}
}
\value{
A ggplot object.
}
\description{
Make a Parity rule 2 (PR2) plot, where the AT-bias [A3/(A3 +T3)] at the third codon position of the four-codon amino acids of entire genes is the ordinate and the GC-bias [G3/(G3 +C3)] is the abscissa. The center of the plot, where both coordinates are 0.5, is where A = U and G = C (PR2), with no bias between the influence of the mutation and selection rates.
}
\details{
For more information about PR2 plot \href{https://www.tandfonline.com/doi/full/10.1038/emi.2016.106}{Butt et al., 2016}.
}
\examples{
\dontrun{
# read DNA from fasta file
df <- read.virus(c("test1.fasta", "test2.fasta", "test3.fasta") , sep = "_")

PR2.plot(df[[1]])
}
}
\author{
Ali Mostafa Anwar \email{aliali.mostafa99@gmail.com} , Salma Bayoumi {salma.ismail.hamed@gmail.com}
}
