% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.boxplot.R
\name{QC.boxplot}
\alias{QC.boxplot}
\title{QC boxplot}
\usage{
QC.boxplot(list.virus)
}
\arguments{
\item{list.virus}{A list of viruses' DNA returned from read.virus function}
}
\value{
A ggplot2 object
}
\description{
boxplot for the amino acid sequence length as a QC
}
\examples{
\dontrun{
# read DNA from fasta file
fasta.virus <- read.virus(c("test1.fasta", "test2.fasta", "test3.fasta"))
QC.boxplot(fasta.virus)
}
}
\author{
Ali Mostafa Anwar \email{aliali.mostafa99@gmail.com} , Salma Bayoumi {salma.ismail.hamed@gmail.com}
}
