% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.cutoff.R
\name{QC.cutoff}
\alias{QC.cutoff}
\title{QC.cutoff}
\usage{
QC.cutoff(list.virus, cut.off.up, cut.off.down)
}
\arguments{
\item{list.virus}{A list of viruses' DNA returned from read.virus function}

\item{cut.off.up}{The maximum threshold of amino acid sequence length}

\item{cut.off.down}{The minimum threshold of amino acid sequence length}
}
\value{
modified list.virus
}
\description{
remove the genes with maximum and minimum threshold amino acid sequence length
}
\examples{
\dontrun{
# read DNA from fasta file
fasta.virus <- read.virus(c("test1.fasta", "test2.fasta", "test3.fasta"))
fasta.virus.mod <- QC.cutoff(fasta.virus)
}
}
\author{
Ali Mostafa Anwar \email{aliali.mostafa99@gmail.com} , Salma Bayoumi {salma.ismail.hamed@gmail.com}
}
