% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCDI.values.R
\name{RCDI.calc}
\alias{RCDI.calc}
\title{Relative Codon Deoptimization Index (RCDI)}
\usage{
RCDI.calc(list.virus, list.host , rscu.host, enc.host, set.length = 5)
}
\arguments{
\item{list.virus}{a list of data frame(s) with virus seq_name and its DNA sequence.}

\item{list.host}{a list of data frame with host seq_name and its DNA sequence.}

\item{rscu.host}{a list of data frame of a hosts' RSCU values.}

\item{enc.host}{a list data frame of a hosts' ENc values.}

\item{set.len}{a number represents a percent that will be used as reference genes from the total host genes.}
}
\value{
A list contains data.frame containing the computed RCDI values for each DNA sequences of each virus.
}
\description{
Measure the Relative Codon Deoptimization Index (RCDI) of DNA sequence.
}
\details{
For more information about RCDI \href{https://bmcresnotes.biomedcentral.com/articles/10.1186/1756-0500-3-87}{Puigbò et al., 2010}
}
\examples{
\dontrun{
# read DNA from fasta file
df.virus <- read.virus(c("test1.fasta", "test2.fasta", "test3.fasta"))
df.host <- read.host(c("host.fasta") , sep = " [locus")
# Calculate RSCU 
rscu.virus <- RSCU.values(fasta.virus)
rscu.host <- RSCU.values(df.host)
enc.novembre.host <- ENc.values.old(df.host)
RCDI.calc(df.virus , df.host, rscu.host, enc.novembre.host, set.length = 100)

}
}
\author{
Ali Mostafa Anwar \email{aliali.mostafa99@gmail.com} , Salma Bayoumi {salma.ismail.hamed@gmail.com}
}
