% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.host.R
\name{read.host}
\alias{read.host}
\title{Read fasta formate and convert it to data frame}
\usage{
read.virus(host.fasta , sep = "|")
}
\arguments{
\item{host.fasta}{a directory path to the host fasta file.}

\item{sep}{for long sequence name separate the fasta header with?}
}
\value{
A list with DNAStringSet.
}
\description{
Read fasta formate and convert it to data frame
}
\examples{
\dontrun{
fastas.host <- read.host(c("File1.fasta", "File2.fasta", "File3.fasta"), sep = "|")
}
}
\author{
Ali Mostafa Anwar \email{aliali.mostafa99@gmail.com} , Salma Bayoumi {salma.ismail.hamed@gmail.com}
}
