% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.virus.R
\name{read.virus}
\alias{read.virus}
\title{Read fasta formate and convert it to data frame}
\usage{
read.virus(virus.fastas, sep = "|")
}
\arguments{
\item{virus.fastas}{a Vector of directory path to the virus fasta file(s).}

\item{sep}{for long sequence name separate the fasta header with?}
}
\value{
A list with DNAStringSet.
}
\description{
Read fasta formate and convert it to data frame
}
\examples{
\dontrun{
fastas.virus <- read.virus(c("File1.fasta", "File2.fasta", "File3.fasta") , sep = "|")
}
}
\author{
Ali Mostafa Anwar \email{aliali.mostafa99@gmail.com} , Salma Bayoumi {salma.ismail.hamed@gmail.com}
}
