% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSCU.Cluster.r
\name{rscu.cluster}
\alias{rscu.cluster}
\title{Relative Synonymous Codon Usage (RSCU) cluster using factoextra package}
\usage{
rscu.cluster(virus.list.rscu, host.rscu, FUNcluster = "kmeans",
         k = 3, hc_metric = "euclidean" , hc_method = "ward.D2", rank = 4,
         codons.exclude = c("ATG", "TAA", "TAG", "TGA", "TGG"))
}
\arguments{
\item{virus.list.rscu}{A list of viruses' RSCU values}

\item{host.rscu}{A list of host RSCU values}

\item{FUNcluster}{a clustering function including "kmeans", "pam", "clara", "fanny", "hclust", "agnes" and "diana". Abbreviation is allowed.}

\item{hc_metric}{character string specifying the metric to be used for calculating dissimilarities between observations. Allowed values are those accepted by the function dist() [including "euclidean", "manhattan", "maximum", "canberra", "binary", "minkowski"] and correlation based distance measures ["pearson", "spearman" or "kendall"]. Used only when FUNcluster is a hierarchical clustering function such as one of "hclust", "agnes" or "diana".}

\item{hc_method}{the agglomeration method to be used (?hclust): "ward.D", "ward.D2", "single", "complete", "average", ...}

\item{rank}{a number specifying the number of PCAs used in the clustering the data, i.e., maximal number of principal components to be used.}

\item{codons.exclude}{exclude codons like start and termination codons}

\item{K}{the number of clusters to be generated. If NULL, the gap statistic is used to estimate the appropriate number of clusters. In the case of kmeans, k can be either the number of clusters, or a set of initial (distinct) cluster centers.}
}
\value{
A ggplot2 object
}
\description{
PCA for the Relative Synonymous Codon Usage (RSCU) of DNA sequence.
}
\examples{
\dontrun{
# read DNA from fasta file
fasta.virus <- read.virus(c("test1.fasta", "test2.fasta", "test3.fasta"))
fasta.host <- read.host(host.fasta)

# Calculate RSCU
rscu.virus <- RSCU.values(fasta.virus)
rscu.host <- RSCU.values(fasta.host)

rscu.cluster(rscu.virus, rscu.host)
}
}
\author{
Ali Mostafa Anwar \email{aliali.mostafa99@gmail.com} , Salma Bayoumi {salma.ismail.hamed@gmail.com}
}
