% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSCU.PCA.r
\name{rscu.pca}
\alias{rscu.pca}
\title{Relative Synonymous Codon Usage (RSCU) PCA}
\usage{
rscu.pca(virus.list.rscu, host.rscu, codons.exclude = c("ATG", "TAA", "TAG", "TGA", "TGG"))
}
\arguments{
\item{virus.list.rscu}{A list of viruses' RSCU values}

\item{host.rscu}{A list of host RSCU values}

\item{codons.exclude}{exclude codons like start and termination codons}
}
\value{
A ggplot2 object
}
\description{
PCA for the Relative Synonymous Codon Usage (RSCU) of DNA sequence.
}
\examples{
\dontrun{
# read DNA from fasta file
fasta.virus <- read.virus(c("test1.fasta", "test2.fasta", "test3.fasta"))
fasta.host <- read.host(host.fasta)

# Calculate RSCU
rscu.virus <- RSCU.values(fasta.virus)
rscu.host <- RSCU.values(fasta.host)

rscu.pca(rscu.virus, rscu.host)
}
}
\author{
Ali Mostafa Anwar \email{aliali.mostafa99@gmail.com} , Salma Bayoumi {salma.ismail.hamed@gmail.com}
}
