% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSCU.values.R
\name{RSCU.values}
\alias{RSCU.values}
\title{Relative Synonymous Codon Usage (RSCU)}
\usage{
RSCU.values(df.fasta)
}
\arguments{
\item{df.fasta}{a data frame with seq_name and its DNA sequence.}
}
\value{
A data.frame containing the computed RSCU values for each codon for each DNA sequences within df.fasta.
}
\description{
Measure the Relative Synonymous Codon Usage (RSCU) of DNA sequence.
}
\details{
For more information about ENc \href{https://academic.oup.com/nar/article-abstract/14/13/5125/1143812?redirectedFrom=fulltext}{Sharp et al., 1986}.
}
\examples{
\dontrun{
# read DNA from fasta file
fasta.virus <- read.virus(c("test1.fasta", "test2.fasta", "test3.fasta"))
# Calculate RSCU 
rscu.virus <- RSCU.values(fasta.virus)
}
}
\author{
Ali Mostafa Anwar \email{aliali.mostafa99@gmail.com} , Salma Bayoumi {salma.ismail.hamed@gmail.com}
}
