% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiD.value.R
\name{SiD.list}
\alias{SiD.list}
\title{Similarity Index (SiD)}
\usage{
SiD.list(RSCU.virus,RSCU.host)
}
\arguments{
\item{RSCU.virus}{a list of data.frame(s) with RSCU a virus codon values.}

\item{RSCU.host}{a list of a data.frame with RSCU a host codon values.}
}
\value{
A data.frame of numeric represent a SiD value.
}
\description{
Measure the Similarity Index (SiD) between a virus and its host codon usage.
}
\details{
For more information about SiD \href{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0077239}{Zhou et al., 2013}.
}
\examples{
\dontrun{
# read DNA from fasta file
fasta.virus <- read.virus(c("test1.fasta", "test2.fasta", "test3.fasta"))
fasta.host <- read.host(c("host.fasta"))
# Calculate RSCU 
rscu.virus <- RSCU.values(fasta.virus)
rscu.host <- RSCU.values(fasta.host)
# Calculate Sid
SiD.list(rscu.virus, rscu.host)
}
}
\author{
Ali Mostafa Anwar \email{aliali.mostafa99@gmail.com} , Salma Bayoumi {salma.ismail.hamed@gmail.com}
}
